Fixed circular imports with AnnotationLabelSerializer by moving to its own script.
